Move server main to workbench and add cycles include
