+++ CMakeLists.txt
@@ -118,114 +118,9 @@ endif()
 # Consolidated examples are built in a single workbench target
 add_subdirectory(examples/workbench)
 
-# Main executable
-add_executable(sep_server src/server_main.cpp)
-# A minimal CLI to test the core engine logic
+# Minimal CLI to test the core engine logic
 add_executable(sep_engine_cli src/main.cpp)
 target_include_directories(sep_engine_cli PRIVATE ${CMAKE_SOURCE_DIR}/include)
-target_link_libraries(sep_engine_cli PRIVATE
-    sep_core
-    sep_memory
-    sep_quantum
-    sep_compat
-)
-
-
-# Additional executable using the same engine core
-add_executable(sep_engine src/server_main.cpp)
-
-# A minimal CLI to test the core engine logic
-target_include_directories(sep_engine_cli
-    PUBLIC
-        ${CMAKE_SOURCE_DIR}/include
-)
-target_link_libraries(sep_engine_cli
-    PRIVATE
-        sep_core
-        sep_memory
-        sep_quantum
-        sep_compat
-)
-
-set_target_properties(sep_server PROPERTIES
-    CUDA_SEPARABLE_COMPILATION ON
-    CUDA_RESOLVE_DEVICE_SYMBOLS ON
-)
-set_target_properties(sep_engine PROPERTIES
-    CUDA_SEPARABLE_COMPILATION ON
-    CUDA_RESOLVE_DEVICE_SYMBOLS ON
-)
-target_include_directories(sep_server
-    PUBLIC
-        ${CMAKE_SOURCE_DIR}/include
-)
-target_include_directories(sep_engine
-    PUBLIC
-        ${CMAKE_SOURCE_DIR}/include
-)
-
-# Minimal CLI executable for testing core engine logic
-target_link_libraries(sep_engine_cli PRIVATE
-    sep_core
-    sep_memory
-    sep_quantum
-    sep_compat
-)
-
-target_link_libraries(sep_server
-    PRIVATE
-        # 1. High-level SEP Engine static libraries
-        sep_api
-        sep_blender
-        sep_audio
-        sep_embeddings
-        sep_quantum
-        sep_memory
-        sep_compat
-        sep_core
-
-        # 2. Cycles and its dependencies
-        cycles_session
-        cycles_scene
-        cycles_device
-        cycles_kernel_osl
-        cycles_subd
-        cycles_util
-        extern_sky
-        extern_cuew
-        extern_hipew
-        extern_sky
-
-        # 3. External dependencies
-        OpenImageIO
-        ${OPENVDB_LIBRARIES}
-        ${OPENSUBDIV_LIBRARIES}
-        openvdb
-        bf_intern_opensubdiv
-        ${OSL_LIBRARIES}
-        ${TBB_LIBRARIES}
-        ${CURL_LIBRARIES}
-        http_parser::http_parser
-        fmt::fmt
-        spdlog::spdlog
-        hiredis
-        pipewire-0.3
-
-        # 4. System libraries
-        ${EMBREE_LIBRARIES}
-        Imath
-        ${OpenPGL_LIBRARIES}
-        ${OpenColorIO_LIBRARIES}
-        ${ALEMBIC_LIBRARIES}
-        ${CMAKE_DL_LIBS}
-        ${Threads_LIBRARIES}
-        ${CUDAToolkit_LIBRARIES}
-        pthread
-        dl
-        rt
-        m
-        z
-)
 
 # ---- Minimal CLI executable ----
 target_link_libraries(sep_engine_cli PRIVATE
@@ -262,10 +157,6 @@ install(TARGETS
     INCLUDES DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
 
-# Install executable separately
-install(TARGETS sep_server sep_engine
-    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
-)
 
 # Install CMake config files
 install(EXPORT sep-targets
